+++ projects/kit/components/textarea/index.ts
@@ -1,4 +1,0 @@
-export * from './deprecated';
-export * from './textarea.component';
-export * from './textarea.directive';
-export * from './textarea.module';